DAL/Repositories/TodoItemRepository.cs
using DAL.Data;
using DAL.Entities;
using DAL.Interfaces;
using Microsoft.EntityFrameworkCore;

namespace DAL.Repositories;

public class TodoItemRepository : ITodoItemRepository
{
    private readonly ApplicationDbContext _context;

    public TodoItemRepository(ApplicationDbContext context)
    {
        _context = context;
    }


    public async Task<IEnumerable<TodoItem>> GetAllAsync()
    {
        return await _context.TodoItems.ToListAsync();
    }

    public async Task AddAsync(TodoItem entity)
    {
        _context.TodoItems.Add(entity);
        await _context.SaveChangesAsync();
    }

    public async Task DeleteByIdAsync(int id)
    {
        var item = await _context.TodoItems.FindAsync(id);
        
        if (item != null)
        {
            _context.TodoItems.Remove(item);
            await _context.SaveChangesAsync();
        }    
    }

    public async Task UpdateAsync(TodoItem entity)
    {
        _context.TodoItems.Update(entity);
        await _context.SaveChangesAsync();    
    }
}